+++ NEWS.md
@@ -2,8 +2,9 @@
 
 # surveydown 0.13.2
 
-- New feature: `sd_validate()` function to validate if question results satisfy certain conditions.
-- Enhanced: Now for the judging conditions in the 3 conditional check functions (`sd_skip_forward()`, `sd_show_if()`, and `sd_validate()`), these judging questions are auto-assigned as required questions.
+- New feature: `sd_stop_if()` function to stop the navigate button if certain conditions cannot be met.
+- Deprecated `sd_skip_forward()` function, and reuse `sd_stop_if()`.
+- Enhanced: Now for the judging conditions in the 3 condition check functions (`sd_show_if()`, `sd_skip_if()`, and `sd_stop_if()`), these judging questions are auto-assigned as required questions.
 
 # surveydown 0.13.1